requirements.txt
streamlit
web3
python-dotenv
eth-account
py-solc-x
requests>=2.31.0
matplotlib==3.9.2
numpy==1.26.4
reportlab==4.1.0
